test: extract inline lib.rs tests to tests/lino_rates_loader_tests.rs
Restore src/lib.rs to under the 1000-line repository limit by moving
the single #[cfg(test)] mod tests block (which exercised
load_rates_from_consolidated_lino) into an outside-the-crate
integration test file. The test uses only public APIs so the move
is behaviour-preserving.

Fixes the Lint and Format Check failure introduced by the public
surface expansion in the previous commit.
