+++ src/lib.rs
@@ -995,32 +995,3 @@ impl Calculator {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_lino_rates_used_in_historical_conversion() {
-        let mut calc = Calculator::new();
-        let content = "conversion:
-  from USD
-  to RUB
-  source 'cbr.ru (Central Bank of Russia)'
-  rates:
-    2021-02-08 74.2602
-    2021-02-09 74.1192";
-
-        let loaded = calc.load_rates_from_consolidated_lino(content);
-        assert!(loaded > 0);
-
-        let date = types::DateTime::from_date(chrono::NaiveDate::from_ymd_opt(2021, 2, 8).unwrap());
-        let rate = calc
-            .parser
-            .currency_db()
-            .get_historical_rate("USD", "RUB", &date);
-        assert!(rate.is_some());
-        let rate_value = rate.unwrap();
-        assert!((rate_value - 74.2602).abs() < 0.001);
-    }
-}
